Frontend: Refactor Project Generator tab
