+++ src/main/webapp/app/springboot/primary/project-generator/index.ts
@@ -1,0 +1,4 @@
+import ProjectGeneratorComponent from './ProjectGenerator.component';
+import ProjectGeneratorVue from './ProjectGenerator.vue';
+
+export { ProjectGeneratorComponent, ProjectGeneratorVue };
